requirements.txt
python-telegram-bot
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
google-generativeai
elevenlabs
python-dotenv
pydantic
